refactor(services): move llm and embedding helpers into injectable services
Address review feedback that the LLM helpers lived outside DI. Replace the
free functions in graphs/llm.ts with two injectable services wired through
DI everywhere; no behaviour change (prompts and model calls are identical).

- add LlmService (render, invoke, defaultFileContent) and EmbeddingService
  (embedText, embedTexts), registered in the component's services
- inject them into every node/service/visualizer/provider that used the
  helpers and call through this.llmService / this.embedder (field named
  llmService to avoid colliding with the existing this.llm model field)
- back the vector stores with the injected EmbeddingService; make the
  in-memory store's cosineSimilarity/matchesObjectFilter and the pg store's
  toVectorLiteral private static methods
- delete graphs/llm.ts and drop its barrel export
- update unit tests to supply the services to the classes under test
