+++ src/components/db-query/nodes/generate-description.node.ts
@@ -1,12 +1,7 @@
 import {inject, service} from '@loopback/core';
 import {graphNode} from '../../../decorators';
-import {
-  IGraphNode,
-  invokeModel,
-  LLMStreamEventType,
-  renderPrompt,
-  RunnableConfig,
-} from '../../../graphs';
+import {IGraphNode, LLMStreamEventType, RunnableConfig} from '../../../graphs';
+import {LlmService} from '../../../services/llm.service';
 import {AiIntegrationBindings} from '../../../keys';
 import {LLMProvider} from '../../../types';
 import {getTextContent} from '../../../utils';
@@ -19,6 +14,8 @@ import {DbQueryConfig} from '../types';
 @graphNode(DbQueryNodes.GenerateDescription)
 export class GenerateDescriptionNode implements IGraphNode<DbQueryState> {
   constructor(
+    @service(LlmService)
+    private readonly llmService: LlmService,
     @inject(AiIntegrationBindings.CheapLLM)
     private readonly llm: LLMProvider,
     @inject(DbQueryAIExtensionBindings.Config)
@@ -74,9 +71,9 @@ Return a short bulleted list where each bullet is one condition, filter, or piec
       data: 'Generating query description.',
     });
 
-    const result = await invokeModel(
+    const result = await this.llmService.invoke(
       this.llm,
-      renderPrompt(this.prompt, {
+      this.llmService.render(this.prompt, {
         prompt: state.prompt,
         sql: state.sql,
         schema: this.schemaHelper.asString(state.schema),
