--- src/graphs/llm.ts
@@ -1,274 +1,0 @@
-/**
- * Internal LLM helper.
- *
- * Replaces LangChain's `PromptTemplate` / `RunnableSequence` / `llm.invoke` /
- * `llm.bindTools` with thin wrappers over the Vercel AI SDK (`generateText`,
- * `embed`, `embedMany`). Messages are native AI SDK `ModelMessage` objects.
- */
-import {randomUUID} from 'node:crypto';
-import {
-  embed,
-  embedMany,
-  generateText,
-  tool as aiTool,
-  type AssistantContent,
-  type EmbeddingModel,
-  type LanguageModel,
-  type ModelMessage,
-  type ToolSet,
-} from 'ai';
-import {AnyObject} from '@loopback/repository';
-import {LLMProvider} from '../types';
-import {Messages} from './messages';
-import {GraphTool, LLMCallbacks, LLMEndResult, RunnableConfig} from './types';
-
-/** An AI SDK model augmented with the optional file builder + default settings. */
-export type ConfiguredModel = LLMProvider;
-
-/** Coerces a template variable to its rendered string (null/undefined → ''). */
-function stringifyVar(value: unknown): string {
-  if (value === undefined || value === null) {
-    return '';
-  }
-  if (typeof value === 'string') {
-    return value;
-  }
-  if (
-    typeof value === 'number' ||
-    typeof value === 'boolean' ||
-    typeof value === 'bigint'
-  ) {
-    return String(value);
-  }
-  // Objects/arrays/other — serialize rather than rely on default toString.
-  return JSON.stringify(value) ?? '';
-}
-
-/** Resolves a `{key}` placeholder, throwing on an unknown variable. */
-function resolveVar(key: string, vars: Record<string, unknown>): string {
-  if (!(key in vars)) {
-    throw new Error(`Missing value for input variable "${key}"`);
-  }
-  return stringifyVar(vars[key]);
-}
-
-/**
- * Renders an f-string template with the exact `{var}` / `{{` / `}}` semantics of
- * `PromptTemplate.fromTemplate`, so rendered prompt text is byte-identical.
- * Single pass so escaped `{{`/`}}` (literal braces, e.g. JSON examples in a
- * prompt) are never mistaken for variables.
- */
-export function renderPrompt(
-  template: string,
-  vars: Record<string, unknown> = {},
-): string {
-  let result = '';
-  let i = 0;
-  while (i < template.length) {
-    const pair = template.slice(i, i + 2);
-    if (pair === '{{') {
-      result += '{';
-      i += 2;
-    } else if (pair === '}}') {
-      result += '}';
-      i += 2;
-    } else if (template[i] === '{') {
-      const end = template.indexOf('}', i + 1);
-      if (end === -1) {
-        result += template[i];
-        i += 1;
-      } else {
-        result += resolveVar(template.slice(i + 1, end).trim(), vars);
-        i = end + 1;
-      }
-    } else {
-      result += template[i];
-      i += 1;
-    }
-  }
-  return result;
-}
-
-function toToolSet(tools: GraphTool[]): ToolSet {
-  const set: ToolSet = {};
-  for (const t of tools) {
-    // No `execute`: the model returns the tool call without running it, so the
-    // graph keeps control of the tool loop (matching the old bindTools flow).
-    set[t.name] = aiTool({
-      description: t.description,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      inputSchema: t.schema as any,
-    });
-  }
-  return set;
-}
-
-export interface InvokeOptions {
-  system?: string;
-  tools?: GraphTool[];
-  temperature?: number;
-  config?: RunnableConfig;
-}
-
-function getModelId(model: LanguageModel): string {
-  if (typeof model === 'string') {
-    return model;
-  }
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return (model as any).modelId ?? 'unknown';
-}
-
-/**
- * Hoists system-role content into a single `system` string (the AI SDK expects
- * system content out of `messages`), returning the remaining non-system turns.
- */
-function hoistSystem(
-  all: ModelMessage[],
-  optionSystem?: string,
-): {system?: string; messages: ModelMessage[]} {
-  const systemParts: string[] = [];
-  const messages: ModelMessage[] = [];
-  for (const message of all) {
-    if (message.role === 'system') {
-      systemParts.push(
-        typeof message.content === 'string' ? message.content : '',
-      );
-    } else {
-      messages.push(message);
-    }
-  }
-  if (optionSystem) {
-    systemParts.unshift(optionSystem);
-  }
-  return {
-    system: systemParts.filter(Boolean).join('\n\n') || undefined,
-    messages,
-  };
-}
-
-/** Builds the `LLMEndResult` shape the token-counter / Langfuse callbacks expect. */
-function toEndResult(usage: {
-  inputTokens?: number;
-  outputTokens?: number;
-}): LLMEndResult {
-  return {
-    generations: [
-      [
-        {
-          message: {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            usage_metadata: {
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              input_tokens: usage.inputTokens ?? 0,
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              output_tokens: usage.outputTokens ?? 0,
-            },
-          },
-        },
-      ],
-    ],
-  };
-}
-
-/** Converts an AI SDK generate result into an assistant message content value. */
-function toAssistantContent(result: {
-  text?: string;
-  toolCalls?: Array<{toolCallId: string; toolName: string; input: unknown}>;
-}): AssistantContent {
-  const toolCalls = result.toolCalls ?? [];
-  if (toolCalls.length === 0) {
-    return result.text ?? '';
-  }
-  return [
-    ...(result.text ? [{type: 'text' as const, text: result.text}] : []),
-    ...toolCalls.map(tc => ({
-      type: 'tool-call' as const,
-      toolCallId: tc.toolCallId,
-      toolName: tc.toolName,
-      input: tc.input,
-    })),
-  ];
-}
-
-function fireCallbacks(
-  callbacks: LLMCallbacks[],
-  fn: (cb: LLMCallbacks) => void,
-): void {
-  for (const cb of callbacks) {
-    fn(cb);
-  }
-}
-
-/**
- * Calls the model once and returns the assistant `ModelMessage` (with any
- * tool-call parts). Fires the run's `handleLLMStart`/`handleLLMEnd` callbacks so
- * token accounting keeps working. Replaces `llm.invoke` /
- * `llm.bindTools(tools).invoke(...)`.
- */
-export async function invokeModel(
-  model: ConfiguredModel,
-  input: string | Messages,
-  options: InvokeOptions = {},
-): Promise<ModelMessage> {
-  const all: ModelMessage[] =
-    typeof input === 'string' ? [{role: 'user', content: input}] : input;
-  const {system, messages} = hoistSystem(all, options.system);
-
-  const runId = randomUUID();
-  const modelName = getModelId(model);
-  const callbacks = options.config?.callbacks ?? [];
-  fireCallbacks(callbacks, cb =>
-    cb.handleLLMStart?.(runId, modelName, {system, messages}),
-  );
-
-  const toolSet = options.tools?.length ? toToolSet(options.tools) : undefined;
-
-  const result = await generateText({
-    ...model.defaultSettings,
-    model,
-    ...(system ? {system} : {}),
-    messages,
-    ...(toolSet ? {tools: toolSet} : {}),
-    ...(options.temperature !== undefined
-      ? {temperature: options.temperature}
-      : {}),
-    abortSignal: options.config?.signal,
-  });
-
-  const content = toAssistantContent(result);
-  const endResult = toEndResult(
-    result.usage ?? {inputTokens: 0, outputTokens: 0},
-  );
-  fireCallbacks(callbacks, cb =>
-    cb.handleLLMEnd?.(runId, endResult, {text: result.text, content}),
-  );
-
-  return {role: 'assistant', content};
-}
-
-/** Embeds a single string. Replaces LangChain embeddings `.embedQuery`. */
-export async function embedText(
-  model: EmbeddingModel,
-  value: string,
-): Promise<number[]> {
-  const {embedding} = await embed({model, value});
-  return embedding as number[];
-}
-
-/** Embeds many strings. Replaces LangChain embeddings `.embedDocuments`. */
-export async function embedTexts(
-  model: EmbeddingModel,
-  values: string[],
-): Promise<number[][]> {
-  const {embeddings} = await embedMany({model, values});
-  return embeddings as number[][];
-}
-
-/** A default file content block for a model that has no custom `getFile`. */
-export function defaultFileContent(file: Express.Multer.File): AnyObject {
-  return {
-    type: 'file',
-    mediaType: 'application/pdf',
-    data: file.buffer?.toString('base64') ?? '',
-  };
-}
