+++ src/services/index.ts
@@ -1,3 +1,5 @@
 export * from './generation.service';
 export * from './token-counter.service';
 export * from './limit-strategies';
+export * from './llm.service';
+export * from './embedding.service';
