+++ src/sub-modules/db/postgresql/vector-store/pgvector.store.ts
@@ -3,17 +3,21 @@ import {
   inject,
   injectable,
   Provider,
+  service,
   ValueOrPromise,
 } from '@loopback/core';
 import * as pg from 'pg';
 import {EmbeddingProvider} from '../../../../types';
 import {AiIntegrationBindings} from '../../../../keys';
 import {juggler} from '@loopback/repository';
+import {EmbeddingService} from '../../../../services/embedding.service';
 import {PgVectorStoreImpl, VectorStore} from '../../../../vector';
 
 @injectable({scope: BindingScope.SINGLETON})
 export class PgVectorStore implements Provider<VectorStore> {
   constructor(
+    @service(EmbeddingService)
+    private readonly embedder: EmbeddingService,
     @inject(AiIntegrationBindings.EmbeddingModel)
     private readonly embeddingModel: EmbeddingProvider,
     @inject(`datasources.writerdb`)
@@ -33,7 +37,7 @@ export class PgVectorStore implements Provider<VectorStore> {
     const reusablePool = this.pgDataSource.connector?.pg as pg.Pool;
     const dsConfig = this.pgDataSource.connector?.settings;
 
-    return new PgVectorStoreImpl(this.embeddingModel, {
+    return new PgVectorStoreImpl(this.embedder, this.embeddingModel, {
       pool: reusablePool,
       schemaName: dsConfig.schema || 'public',
       tableName: 'semantic_cache',
